requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
beautifulsoup4==4.12.3
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
deep-translator==1.11.4
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
fsspec==2024.10.0
fugashi==1.3.3
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
ipython==8.29.0
jaconv==0.4.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
loguru==0.7.2
manga-ocr==0.1.13
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mojimoji==0.0.13
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
ninja==1.11.1.1
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==11.0.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyperclip==1.9.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
sacremoses==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
termcolor==2.5.0
thinc==8.3.2
tifffile==2024.9.20
tokenizers==0.20.2
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
unidic-lite==1.0.8
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
win32-setctime==1.1.0
wrapt==1.16.0
